+++ isodata/pjm.py
@@ -1,14 +1,22 @@
+from ast import Raise
+
 import pandas as pd
 import requests
 
 import isodata
-from isodata.base import FuelMix, ISOBase
+from isodata.base import FuelMix, ISOBase, Markets
 
 
 class PJM(ISOBase):
     name = "PJM"
     iso_id = "pjm"
     default_timezone = "US/Eastern"
+
+    markets = [
+        Markets.REAL_TIME_5_MIN,
+        Markets.REAL_TIME_HOURLY,
+        Markets.DAY_AHEAD_HOURLY,
+    ]
 
     def get_latest_fuel_mix(self):
         mix = self.get_fuel_mix_today()
@@ -150,6 +158,123 @@ class PJM(ISOBase):
     # todo https://dataminer2.pjm.com/feed/load_frcstd_hist/definition
     # def get_historical_forecast(self, date):
     # pass
+
+    def get_pnode_ids(self):
+        data = {
+            "startRow": 1,
+            "rowCount": 500000,
+            "fields": "effective_date,pnode_id,pnode_name,pnode_subtype,pnode_type,termination_date,voltage_level,zone",
+        }
+        r = self._get_pjm_json("pnode", params=data)
+
+        data = pd.DataFrame(r["items"])
+        return data
+
+    def get_latest_lmp(self, market: str, locations: list = None):
+        """Currently only supports DAY_AHEAD_HOURlY"""
+        market = Markets(market)
+        if market != Markets.DAY_AHEAD_HOURLY:
+            raise NotImplementedError("Only supports DAY_AHEAD_HOURLY")
+        return self._latest_lmp_from_today(market, locations)
+
+    def get_lmp_today(self, market: str, locations: list = None):
+        """Get lmp for today
+        Currently only supports DAY_AHEAD_HOURlY
+        """
+        # TODO try to find a different source of data for real time
+        market = Markets(market)
+        if market != Markets.DAY_AHEAD_HOURLY:
+            raise NotImplementedError("Only supports DAY_AHEAD_HOURLY")
+        return self._today_from_historical(self.get_historical_lmp, market, locations)
+
+    def get_historical_lmp(self, date, market: str, locations: list = None):
+        """Returns LMP at a previous date
+
+        Args:
+            date (str or datetime.date): date to get LMPs for
+            market (str):  Supported Markets: REAL_TIME_5_MIN, REAL_TIME_HOURLY, DAY_AHEAD_HOURlY
+            locations (list, optional):  list of pnodeid to get LMPs for. Defaults to Hubs. Use get_pnode_ids() to get a list of posssible pnodeids
+
+        """
+        date = date = isodata.utils._handle_date(date)
+        tomorrow = date + pd.DateOffset(1)
+
+        if locations is None:
+            locations = [
+                "51217",
+                "116013751",
+                "35010337",
+                "34497151",
+                "34497127",
+                "34497125",
+                "33092315",
+                "33092313",
+                "33092311",
+                "4669664",
+                "51288",
+                "51287",
+            ]
+
+        market = Markets(market)
+        if market == Markets.REAL_TIME_5_MIN:
+            market_endpoint = "rt_fivemin_hrl_lmps"
+            market_type = "rt"
+        elif market == Markets.REAL_TIME_HOURLY:
+            market_endpoint = "rt_hrl_lmps"
+            market_type = "rt"
+        elif market == Markets.DAY_AHEAD_HOURLY:
+            market_endpoint = "da_hrl_lmps"
+            market_type = "da"
+        else:
+            raise ValueError(
+                "market must be one of REAL_TIME_5_MIN, REAL_TIME_HOURLY, DAY_AHEAD_HOURLY",
+            )
+
+        #  TODO implement paging since row count can exceed 1000000
+        data = {
+            "datetime_beginning_ept": date.strftime("%m/%d/%Y 00:00")
+            + "to"
+            + tomorrow.strftime("%m/%d/%Y 00:00"),
+            "startRow": 1,
+            "rowCount": 1000000,
+            "fields": f"congestion_price_{market_type},datetime_beginning_ept,datetime_beginning_utc,equipment,marginal_loss_price_{market_type},pnode_id,pnode_name,row_is_current,system_energy_price_{market_type},total_lmp_{market_type},type,version_nbr,voltage,zone",
+            "pnode_id": ";".join(locations),
+        }
+
+        r = self._get_pjm_json(market_endpoint, params=data)
+        data = pd.DataFrame(r["items"]).rename(
+            columns={
+                "datetime_beginning_ept": "Time",
+                "pnode_id": "Location",
+                "pnode_name": "Location Name",
+                "type": "Location Type",
+                f"total_lmp_{market_type}": "LMP",
+                f"system_energy_price_{market_type}": "Energy",
+                f"congestion_price_{market_type}": "Congestion",
+                f"marginal_loss_price_{market_type}": "Loss",
+            },
+        )
+
+        data["Market"] = market.value
+
+        data = data[
+            [
+                "Time",
+                "Market",
+                "Location",
+                "Location Name",
+                "Location Type",
+                "LMP",
+                "Energy",
+                "Congestion",
+                "Loss",
+            ]
+        ]
+
+        return data
+        import pdb
+
+        pdb.set_trace()
 
     def _get_pjm_json(self, endpoint, params):
         r = self._get_json(