+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## vNext
 
 - Add notes to Ercot status
+- Add `.status_homepage` url to ISOs that report a status
 
 ## v0.7.0 - Aug 23, 2022