+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Add notes to Ercot status
 - Add `.status_homepage` url to ISOs that report a status
-- Add historical RTM Settlement Prices to Ercot
+- Add Ercot Historical RTM Settlement Point Prices (SPPs)
 
 ## v0.7.0 - Aug 23, 2022