+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNext
 
+- PJM: add lmp prices for 3 markets: real time 5 minutes, real time hourly, day ahead hourly
 - Add notes to Ercot status
 - Add `.status_homepage` url to ISOs that report a status
 - Add Ercot Historical RTM Settlement Point Prices (SPPs)